requirements.txt
# FastAPI e servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Banco de dados
sqlalchemy==2.0.25
alembic==1.13.1

# Telegram Bot
python-telegram-bot==20.7

# Claude IA
anthropic==0.18.1

# Utilidades
python-dotenv==1.0.0
python-multipart==0.0.6
httpx~=0.25.2

# Data e hora
python-dateutil==2.8.2
pytz==2023.3

# Validação e parsing
email-validator==2.1.0

# Testes
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Desenvolvimento
black==24.1.1
flake8==7.0.0
mypy==1.8.0